feat(script): add admin check to ensure the script runs with elevated privileges
- Added a check to validate if the script is running as an administrator.
- If not, the script warns the user and terminates.

+++ optimize_explorer_perfomance.ps1
@@ -1,3 +1,9 @@
+# Ensure the script can run with elevated privileges
+if (-NOT ([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] "Administrator")) {
+    Write-Warning "Please run this script as an Administrator!"
+    exit
+}
+
 # Eliminar Bags y BagMRU si existen / Delete Bags and BagMRU if they exist
 Remove-Item -Path "HKCU:\SOFTWARE\Classes\Local Settings\Software\Microsoft\Windows\Shell\Bags" -Recurse -Force -ErrorAction SilentlyContinue
 Remove-Item -Path "HKCU:\SOFTWARE\Classes\Local Settings\Software\Microsoft\Windows\Shell\BagMRU" -Recurse -Force -ErrorAction SilentlyContinue
